If a read or write fails (as opposed to a read returning 0 to indicate EOF) then the connection is *dead*, aka. status == -1, not merely half-closed (status == 0).
Prior to this commit, if a connection was reset then spiped would fail to
close the other side, resulting in sockets and connections to the target
being held open.  In the case of ssh-over-spiped, this could result in
"zombie" login sessions lingering indefinitely if ssh protocol keepalive
messages are not enabled.


git-svn-id: http://spiped.googlecode.com/svn/trunk@72 592f418a-0512-ad62-f63d-38f288cbaac7